+++ zarrs/src/array/array_ops/array_read_ops_array_cached.rs
@@ -1,6 +1,7 @@
 use inherent::inherent;
 use std::sync::Arc;
 use unsafe_cell_slice::UnsafeCellSlice;
+use zarrs_codec::EncodedSubchunk;
 
 #[cfg(not(target_arch = "wasm32"))]
 use rayon::iter::ParallelIterator;
@@ -363,14 +364,21 @@ where
     pub fn retrieve_encoded_subchunk(
         &self,
         subchunk_indices: &[u64],
-    ) -> Result<Option<Vec<u8>>, ArrayError>;
+    ) -> Result<Option<EncodedSubchunk<'static>>, ArrayError>;
 
     #[allow(clippy::missing_errors_doc)]
     pub fn retrieve_encoded_subchunk_at_level(
         &self,
         level: usize,
         subchunk_indices: &[u64],
-    ) -> Result<Option<Vec<u8>>, ArrayError>;
+    ) -> Result<Option<EncodedSubchunk<'static>>, ArrayError>;
+
+    #[allow(clippy::missing_errors_doc)]
+    pub fn encoded_subchunk_shape_at_level(
+        &self,
+        level: usize,
+        subchunk_indices: &[u64],
+    ) -> Result<ChunkShape, ArrayError>;
 
     #[allow(clippy::missing_errors_doc)]
     pub fn retrieve_subchunk<T: FromArrayBytes>(