+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_decoder_sync.rs
@@ -207,7 +207,7 @@ impl ArrayPartialDecoderSubchunkingTraits for ShardingPartialDecoder {
         nested_subchunk_codecs(&self.inner_codecs)
     }
 
-    fn retrieve_encoded_subchunk(
+    fn retrieve_encoded_subchunk_bytes(
         &self,
         subchunk_indices: &[u64],
         options: &CodecOptions,