+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_encoder_async.rs
@@ -128,7 +128,7 @@ impl AsyncArrayPartialDecoderSubchunkingTraits for AsyncShardingPartialEncoder {
         nested_subchunk_codecs(&self.inner_codecs)
     }
 
-    async fn retrieve_encoded_subchunk<'a>(
+    async fn retrieve_encoded_subchunk_bytes<'a>(
         &'a self,
         subchunk_indices: &[u64],
         options: &CodecOptions,